refactor(#8): simplify trackers and trackers settings ✨
