--- issue-solver/src/issue_solver/issue_trackers/azure_devops/azure_devops_issue_tracker.py
@@ -1,12 +1,0 @@
-from issue_solver.issue_trackers.issue_tracker import (
-    IssueTracker,
-    IssueInfo,
-    IssueReference,
-)
-
-
-class AzureDevOpsIssueTracker(IssueTracker):
-    def describe_issue(self, issue_reference: IssueReference) -> IssueInfo | None:
-        raise NotImplementedError(
-            "AzureDevOpsIssueTracker.get_issue_description is not implemented"
-        )
