--- issue-solver/src/issue_solver/issue_trackers/http_based/http_based_issue_tracker.py
@@ -1,10 +1,0 @@
-from issue_solver.issue_trackers.issue_tracker import (
-    IssueTracker,
-    IssueInfo,
-    IssueReference,
-)
-
-
-class HttpBasedIssueTracker(IssueTracker):
-    def describe_issue(self, issue_reference: IssueReference) -> IssueInfo | None:
-        pass
