--- issue-solver/src/issue_solver/issue_trackers/jira/jira_issue_tracker.py
@@ -1,12 +1,0 @@
-from issue_solver.issue_trackers.issue_tracker import (
-    IssueTracker,
-    IssueInfo,
-    IssueReference,
-)
-
-
-class JiraIssueTracker(IssueTracker):
-    def describe_issue(self, issue_reference: IssueReference) -> IssueInfo | None:
-        raise NotImplementedError(
-            "JiraIssueTracker.get_issue_description is not implemented"
-        )
