+++ issue-solver/src/issue_solver/issues/trackers/gitlab_issue_tracker.py
@@ -1,17 +1,26 @@
+from enum import StrEnum
 from http import HTTPStatus
 from pathlib import Path
-from typing import Self, assert_never
+from typing import Self, assert_never, Literal
 
 import gitlab
+from pydantic import AnyUrl, Field
+from pydantic_core import Url
+from pydantic_settings import SettingsConfigDict
 
-from issue_solver.issue_trackers.gitlab.settings import GitlabIssueTrackerSettings
-from issue_solver.issue_trackers.issue_tracker import (
+from issue_solver.issues.trackers.issue_tracker import (
     IssueTracker,
     IssueInfo,
     IssueReference,
     IssueInternalId,
     IssueId,
 )
+from issue_solver.issues.settings import ApiBasedIssueTrackerSettings
+
+
+class GitlabObjectType(StrEnum):
+    ISSUE = "ISSUE"
+    MR = "MR"
 
 
 class GitlabIssueTracker(IssueTracker):
@@ -70,7 +79,7 @@ class GitlabIssueTracker(IssueTracker):
     @classmethod
     def of(
         cls,
-        settings: GitlabIssueTrackerSettings,
+        settings: "GitlabIssueTracker.Settings",
     ) -> Self:
         gitlab_client = gitlab.Gitlab(
             url=str(settings.base_url),
@@ -78,3 +87,20 @@ class GitlabIssueTracker(IssueTracker):
             api_version=settings.api_version,
         )
         return cls(gitlab_client)
+
+    class Settings(ApiBasedIssueTrackerSettings):
+        type: Literal["GITLAB"] = "GITLAB"
+        base_url: AnyUrl = Field(
+            description="Base URL for the GitLab.",
+            default=Url("https://gitlab.com"),
+        )
+        api_version: str = Field(
+            description="API version for the issue tracker.", default="4"
+        )
+        object_type: GitlabObjectType = Field(
+            default=GitlabObjectType.ISSUE,
+            description="Determines whether we retrieve from an ISSUE or a MR in GitLab.",
+        )
+        model_config = SettingsConfigDict(
+            env_prefix="GITLAB_", env_file=".env", env_file_encoding="utf-8"
+        )
