+++ issue-solver/src/issue_solver/issues/trackers/jira_issue_tracker.py
@@ -1,0 +1,33 @@
+from typing import Literal
+
+from pydantic import AnyUrl, Field
+from pydantic_core import Url
+from pydantic_settings import SettingsConfigDict
+
+from issue_solver.issues.trackers.issue_tracker import (
+    IssueTracker,
+    IssueInfo,
+    IssueReference,
+)
+from issue_solver.issues.settings import ApiBasedIssueTrackerSettings
+
+
+class JiraIssueTracker(IssueTracker):
+    def describe_issue(self, issue_reference: IssueReference) -> IssueInfo | None:
+        raise NotImplementedError(
+            "JiraIssueTracker.get_issue_description is not implemented"
+        )
+
+    class Settings(ApiBasedIssueTrackerSettings):
+        type: Literal["JIRA"] = "JIRA"
+        base_url: AnyUrl = Field(
+            description="Base URL for the Jira.",
+            default=Url("https://jira.atlassian.com"),
+        )
+        project_id: str = Field(description="ID of the project in the issue tracker.")
+
+        model_config = SettingsConfigDict(
+            env_prefix="JIRA_",
+            env_file=".env",
+            env_file_encoding="utf-8",
+        )
